fix: 경로 수정및 폴더 구조 변경
