+++ movie_project/src/App.js
@@ -1,25 +1,25 @@
-import logo from './logo.svg';
-import './App.css';
+import { Routes, Route } from "react-router-dom";
+import Login from "./Login";
+import Home from "./Pages/Home";
+import SearchResult from "./Pages/SearchResult";
+import MyPage from "./Pages/MyPage";
+import NotFound from "./Pages/NotFound";
+import Layout from "./LayOut";
+import "bootstrap/dist/css/bootstrap.min.css";
 
-function App() {
+const App = () => {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
+    <Routes>
+      <Route path="/" element={<Login />}></Route>
+      <Route path="/login/" element={<Layout />}>
+        <Route index element={<Home />} />
+      </Route>
+      <Route path="/search/" element={<SearchResult />} />
+      {/* SearchResult결과는 section3 컴포넌트 이용 */}
+      <Route path="/mypage" element={<MyPage />} />
+      <Route path="*" element={<NotFound />} />
+    </Routes>
   );
-}
+};
 
 export default App;
