+++ movie_project/src/Footer.js
