+++ movie_project/src/Pages/SearchResult.js
